4.27更新到3.3.6，测试过基本功能可用，未严格测试

+++ Source/LGUI/Public/Event/LGUIPointerEventData.h
@@ -105,13 +105,13 @@ public:
 		ULGUIBaseRaycaster* pressRaycaster;
 	/** the last time when trigger click(get time from GetWorld()->TimeSeconds), can be used to tell double click */
 	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = "LGUI")
-		double clickTime;
+		float clickTime;
 	/** the last time when trigger release(get time from GetWorld()->TimeSeconds). */
 	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = "LGUI")
-		double releaseTime;
+		float releaseTime;
 	/** the last time when trigger press(get time from GetWorld()->TimeSeconds). */
 	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = "LGUI")
-		double pressTime;
+		float pressTime;
 
 	/** is dragging? */
 	UPROPERTY(EditAnywhere, BlueprintReadOnly, Category = "LGUI")
